requirements.txt
Django==4.2.1
djangorestframework
pyyaml
requests
django-cors-headers